\name{SimpactUtility-package}
\alias{SimpactUtility-package}
\alias{SimpactUtility}
\docType{package}
\title{
Utility functions for Simpact Cyan
}
\description{
Utility functions for Simpact Cyan
}
\details{
\tabular{ll}{
Package: \tab SimpactUtility\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-13\cr
License: \tab What license is it under?\cr
}
Simpac Cyan Utility
}
\author{
Fei MENG
Maintainer: Fei MENG <fei.d.meng@gmail.com>
}
\references{

}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{}}
}
\examples{
}
